added javadoc comments and deleted old classes
