+++ de.tu_bs.cs.isf.e4cf.text_editor/src/de/tu_bs/cs/isf/e4cf/text_editor/stringtable/EditorST.java
@@ -27,6 +27,7 @@ public class EditorST {
 	public static final String EXTP_INDENT = "indentation";
 	public static final String EXTP_FORMAT = "formatting";
 	public static final String EXTP_EXTENSION = "file_extension";
+	public static final String EXTP_STYLE = "css";
 	
 	//icons
 	public static final String JAVA_ICON = "icons/java-icon.png";
